Change calculation and handling changes of selecteddates to app instead of calendardates. This is better not that calendar automatically closes after date selections
