+++ client/components/CalendarDates.jsx
@@ -52,6 +52,7 @@ const CalendarDates = ({
   selectedEndDate,
   changeSelectedDates,
   changeStartDateView,
+  changeViewCalendar,
 }) => {
   const firstDate = new Date(currentYear, currentMonth, 1);
   const lastDate = new Date(currentYear, currentMonth + 1, 0);
@@ -64,39 +65,39 @@ const CalendarDates = ({
 
   // update selectedDates state based on selected start date and selected end date.
   // If statements are used to make sure there is not an infinite loop.
-  if (selectedStartDate && selectedEndDate && selectedDates.length === 0) {
-    if (selectedStartDate.year <= selectedEndDate.year) {
-      if (selectedStartDate.month <= selectedEndDate.month) {
-        if (selectedStartDate.day < selectedEndDate.day) {
-          const startDate = new Date(
-            selectedStartDate.year,
-            selectedStartDate.month,
-            selectedStartDate.day,
-          );
-          const endDate = new Date(
-            selectedEndDate.year,
-            selectedEndDate.month,
-            selectedEndDate.day,
-          );
-          const selectedDays = [];
-          while (startDate.toDateString() !== endDate.toDateString()) {
-            selectedDays.push({
-              year: startDate.getFullYear(),
-              month: startDate.getMonth(),
-              day: startDate.getDate(),
-            });
-            startDate.setDate(startDate.getDate() + 1);
-          }
-          selectedDays.push({
-            year: endDate.getFullYear(),
-            month: endDate.getMonth(),
-            day: endDate.getDate(),
-          });
-          changeSelectedDates(selectedDays);
-        }
-      }
-    }
-  }
+  // if (selectedStartDate && selectedEndDate && selectedDates.length === 0) {
+  //   if (selectedStartDate.year <= selectedEndDate.year) {
+  //     if (selectedStartDate.month <= selectedEndDate.month) {
+  //       if (selectedStartDate.day < selectedEndDate.day) {
+  //         const startDate = new Date(
+  //           selectedStartDate.year,
+  //           selectedStartDate.month,
+  //           selectedStartDate.day,
+  //         );
+  //         const endDate = new Date(
+  //           selectedEndDate.year,
+  //           selectedEndDate.month,
+  //           selectedEndDate.day,
+  //         );
+  //         const selectedDays = [];
+  //         while (startDate.toDateString() !== endDate.toDateString()) {
+  //           selectedDays.push({
+  //             year: startDate.getFullYear(),
+  //             month: startDate.getMonth(),
+  //             day: startDate.getDate(),
+  //           });
+  //           startDate.setDate(startDate.getDate() + 1);
+  //         }
+  //         selectedDays.push({
+  //           year: endDate.getFullYear(),
+  //           month: endDate.getMonth(),
+  //           day: endDate.getDate(),
+  //         });
+  //         changeSelectedDates(selectedDays);
+  //       }
+  //     }
+  //   }
+  // }
 
   // create array of booked days for current month
   for (let i = 0; i < bookedDates.length; i++) {
@@ -139,6 +140,7 @@ const CalendarDates = ({
               changeSelectedStartDate(day, currentMonth, currentYear);
             } else {
               changeSelectedEndDate(day, currentMonth, currentYear);
+              changeViewCalendar(false);
             }
             changeSelectedDates();
           }}
@@ -157,6 +159,7 @@ const CalendarDates = ({
               changeSelectedStartDate(day, currentMonth, currentYear);
             } else {
               changeSelectedEndDate(day, currentMonth, currentYear);
+              changeViewCalendar(false);
             }
             changeSelectedDates();
           }}
@@ -191,6 +194,7 @@ CalendarDates.propTypes = {
   selectedEndDate: PropTypes.any,
   changeSelectedDates: PropTypes.func.isRequired,
   changeStartDateView: PropTypes.func.isRequired,
+  changeViewCalendar: PropTypes.func.isRequired,
 };
 
 CalendarDates.defaultProps = {
